backend/requirements.txt
stability-sdk==0.8.4
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
segment-anything==1.0
onnx-simplifier==0.4.19
onnxruntime==1.15.1
matplotlib==3.8.0
replicate==0.12.0
langchain==0.0.292
opencv-python==4.8.0.76
pandas==2.1.0
requests==2.31.0
tqdm==4.66.1
Pillow==9.3.0
transformers==4.33.1
accelerate==0.23.0
openai==0.28.0
firebase-admin==6.2.0
firebase
google-cloud-storage==2.10.0
uvicorn
fastapi
python-multipart
ibm-watson-machine-learning

